Added Tests for AM-1-T1 and ESCP-4-T2
