+++ backend/.idea/.gitignore
@@ -2,7 +2,7 @@
 /shelf/
 /workspace.xml
 # Datasource local storage ignored files
-/dataSources/
+/../../../../../../../../../../:\Users\Ferran\Documents\Universitat\4\ES\Grup-ES\backend\.idea/dataSources/
 /dataSources.local.xml
 # Editor-based HTTP Client requests
 /httpRequests/
