--- backend/.idea/inspectionProfiles/Project_Default.xml
@@ -1,19 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="ignoredPackages">
-        <value>
-          <list size="6">
-            <item index="0" class="java.lang.String" itemvalue="slidingwindow" />
-            <item index="1" class="java.lang.String" itemvalue="fire" />
-            <item index="2" class="java.lang.String" itemvalue="dill" />
-            <item index="3" class="java.lang.String" itemvalue="tensorpack" />
-            <item index="4" class="java.lang.String" itemvalue="pycocotools" />
-            <item index="5" class="java.lang.String" itemvalue="argparse" />
-          </list>
-        </value>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
